Move JIT-related files to jit dir. Remove IR interpreter
