+++ torch/csrc/Module.cpp
@@ -9,7 +9,8 @@
 #include <ATen/ATen.h>
 
 #include "torch/csrc/utils/python_strings.h"
-#include "torch/csrc/autograd/python_tracer.h"
+#include "torch/csrc/jit/python_tracer.h"
+#include "torch/csrc/jit/python_ir.h"
 
 #ifdef WITH_CUDNN
 #include "cudnn/Module.h"
@@ -551,10 +552,12 @@ extern PyObject * THCPModule_cudaUnlockMutex(PyObject *module);
 
 extern PyObject * THCSPModule_initExtension(PyObject *self);
 #endif
+extern PyObject * THPJIT_initExtension(PyObject *self);
 
 static PyMethodDef TorchMethods[] = {
   {"_initExtension",  (PyCFunction)THPModule_initExtension,   METH_O,       NULL},
   {"_autograd_init",  (PyCFunction)THPAutograd_initExtension, METH_NOARGS,  NULL},
+  {"_jit_init",       (PyCFunction)THPJIT_initExtension,      METH_NOARGS,  NULL},
   {"_tracer_enter",   (PyCFunction)THPTracer_enter,           METH_VARARGS, NULL},
   {"_tracer_exit",    (PyCFunction)THPTracer_exit,            METH_VARARGS, NULL},
   {"_add_docstr",     (PyCFunction)THPModule_addDocStr,       METH_VARARGS, NULL},
