--- torch/csrc/autograd/tracer.cpp
@@ -1,7 +1,0 @@
-#include "torch/csrc/autograd/tracer.h"
-
-namespace torch { namespace autograd {
-
-TracingState GlobalTracingState;
-
-}}
