+++ torch/csrc/jit/python_tracer.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <Python.h>
-#include "torch/csrc/autograd/tracer.h"
+#include "torch/csrc/jit/tracer.h"
 
 PyObject * THPTracer_enter(PyObject *_unused, PyObject* args);
 PyObject * THPTracer_exit(PyObject *_unused, PyObject* args);
