+++ torch/csrc/jit/tracer.cpp
@@ -1,0 +1,7 @@
+#include "torch/csrc/jit/tracer.h"
+
+namespace torch { namespace jit {
+
+TracingState GlobalTracingState;
+
+}}
